Fix the split package between the smallrye-reactive-messaging-jackson and smallrye-reactive-messaging-jsonb modules
