+++ smallrye-reactive-messaging-jackson/src/test/java/io/smallrye/reactive/messaging/json/jackson/TestObject.java
@@ -1,6 +1,8 @@
-package io.smallrye.reactive.messaging.json;
+package io.smallrye.reactive.messaging.json.jackson;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import io.smallrye.reactive.messaging.json.JsonMapping;
 
 /**
  * Test object to use in {@link JsonMapping} implementation.
