+++ smallrye-reactive-messaging-jackson/src/test/java/io/smallrye/reactive/messaging/support/JmsTestBase.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import io.smallrye.config.SmallRyeConfigProviderResolver;
 import io.smallrye.config.inject.ConfigExtension;
 import io.smallrye.reactive.messaging.jms.JmsConnector;
-import io.smallrye.reactive.messaging.json.JacksonMapping;
-import io.smallrye.reactive.messaging.json.ObjectMapperProvider;
+import io.smallrye.reactive.messaging.json.jackson.JacksonMapping;
+import io.smallrye.reactive.messaging.json.jackson.ObjectMapperProvider;
 import io.smallrye.reactive.messaging.providers.MediatorFactory;
 import io.smallrye.reactive.messaging.providers.connectors.ExecutionHolder;
 import io.smallrye.reactive.messaging.providers.connectors.WorkerPoolRegistry;
