+++ smallrye-reactive-messaging-jsonb/src/test/java/io/smallrye/reactive/messaging/json/jsonb/JsonBMappingTest.java
@@ -1,4 +1,4 @@
-package io.smallrye.reactive.messaging.json;
+package io.smallrye.reactive.messaging.json.jsonb;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import io.smallrye.reactive.messaging.support.JmsTestBase;
 import io.smallrye.reactive.messaging.test.common.config.MapBasedConfig;
 
-public class JacksonMappingTest extends JmsTestBase {
+public class JsonBMappingTest extends JmsTestBase {
 
     @Test
     @DisplayName("Test the conversion from string to object and back")
@@ -20,7 +20,7 @@ public class JacksonMappingTest extends JmsTestBase {
         addConfig(config);
         WeldContainer container = deploy();
 
-        JacksonMapping mapping = container.select(JacksonMapping.class).get();
+        JsonBMapping mapping = container.select(JsonBMapping.class).get();
         final String testObjectAsJson = "{\"my_id\": 1, \"my_Payload\": \"Lorem ipsum\"}";
         assertThat(mapping.toJson(mapping.fromJson(testObjectAsJson, TestObject.class))).isNotNull();
     }
