+++ smallrye-reactive-messaging-jsonb/src/test/java/io/smallrye/reactive/messaging/json/jsonb/Person.java
@@ -1,4 +1,4 @@
-package io.smallrye.reactive.messaging.json;
+package io.smallrye.reactive.messaging.json.jsonb;
 
 public class Person {
