+++ smallrye-reactive-messaging-jsonb/src/test/java/io/smallrye/reactive/messaging/json/jsonb/TestObject.java
@@ -1,6 +1,8 @@
-package io.smallrye.reactive.messaging.json;
+package io.smallrye.reactive.messaging.json.jsonb;
 
 import jakarta.json.bind.annotation.JsonbProperty;
+
+import io.smallrye.reactive.messaging.json.JsonMapping;
 
 /**
  * Test object to use in {@link JsonMapping} implementation.
