+++ smallrye-reactive-messaging-jsonb/src/test/java/io/smallrye/reactive/messaging/support/JmsTestBase.java
@@ -9,8 +9,8 @@ import org.junit.jupiter.api.BeforeEach;
 import io.smallrye.config.SmallRyeConfigProviderResolver;
 import io.smallrye.config.inject.ConfigExtension;
 import io.smallrye.reactive.messaging.jms.JmsConnector;
-import io.smallrye.reactive.messaging.json.JsonBMapping;
-import io.smallrye.reactive.messaging.json.JsonBProvider;
+import io.smallrye.reactive.messaging.json.jsonb.JsonBMapping;
+import io.smallrye.reactive.messaging.json.jsonb.JsonBProvider;
 import io.smallrye.reactive.messaging.providers.MediatorFactory;
 import io.smallrye.reactive.messaging.providers.connectors.ExecutionHolder;
 import io.smallrye.reactive.messaging.providers.connectors.WorkerPoolRegistry;
